# docs/developer/openapi/banzami-sandbox.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Banzami Sandbox API",
    "version": "docs-2026-09-14",
    "description": "Machine-readable description of the Banzami Developer API as served by the Public Sandbox — one developer platform, two financial environments. The Sandbox is self-service and uses fictitious value. Financial LIVE is NOT READY: it requires institutional approval and is fail-closed — bz_live_ keys are refused and no Live key can be issued from the Sandbox. PT docs: developers.banzami.com/docs · EN docs: developers.banzami.com/docs/en. Every operation described is served by the deployed Sandbox gateway; an operation a developer key cannot call says so and names the refusal it answers. Each operation names the scope it requires in x-banzami-scope (null for a route that takes no API key). POST /v1/payment-links/{id}/mark-used is retired (410 ROUTE_RETIRED) and is not described. All example values are obvious placeholders (bz_test_sk_XXXX, idem_xxx, payment_session_xxx, bzst_xxx) — never real credentials."
  },
  "servers": [
    {
      "url": "https://sandbox-api.banzami.com",
      "description": "Banzami Sandbox — operational for integration; money is fictitious"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/v1/payment-sessions/{id}/link": {
      "get": {
        "tags": [
          "payment-sessions"
        ],
        "operationId": "getPaymentSessionLink",
        "summary": "Fetch the session's public payment link",
        "x-banzami-scope": "payment_sessions:read",
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Public payment link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "enum": [
                        "PAYMENT_LINK"
                      ]
                    },
                    "slug": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "slug",
                    "url"
                  ]
                },
                "example": {
                  "type": "PAYMENT_LINK",
                  "slug": "slug_xxx",
                  "url": "https://pay.banzami.com/pay/slug_xxx"
                }
              }
            }
          },
          "404": {
            "description": "Session has no payment link (code NO_LINK) or session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/payment-links": {
      "post": {
        "tags": [
          "payment-links"
        ],
        "operationId": "createPaymentLink",
        "summary": "Create a reusable payment link",
        "x-banzami-scope": "payment_links:write",
        "description": "Creates a payment link with a public slug (pay.banzami.com/pay/{slug}). Omit amount_minor for an open-amount link. Credential: merchant credential, or developer key with the payment_links scope on a project with an ACTIVE binding.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentLinkRequest"
              },
              "example": {
                "wallet_id": "wallet_xxx",
                "amount_minor": 25000,
                "currency": "AOA",
                "description": "Order #123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                },
                "example": {
                  "slug": "slug_xxx",
                  "amount_minor": 25000,
                  "currency": "AOA",
                  "description": "Order #123",
                  "status": "ACTIVE",
                  "expires_at": null,
                  "paid_at": null,
                  "merchant_name": "Example Store"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "tags": [
          "payment-links"
        ],
        "operationId": "listPaymentLinks",
        "summary": "List payment links",
        "x-banzami-scope": "payment_links:read",
        "responses": {
          "200": {
            "description": "Payment links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentLink"
                      }
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true,
                      "description": "Present when another page may follow."
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "limit outside 1–100, or a cursor that is not a previous next_cursor (INVALID_PARAM)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Lists the project's payment links, newest first, a page at a time. Pass next_cursor back as cursor for the next page; its absence means the last page.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The next_cursor of the previous page, unchanged."
          }
        ]
      }
    },
...
    "/v1/consumers/handle/{handle}": {
      "get": {
        "tags": [
          "handles"
        ],
        "operationId": "resolveHandle",
        "summary": "Confirm a consumer @banza exists",
        "x-banzami-scope": "customers:read",
        "description": "Resolves a consumer @banza to its handle and display name — enough to confirm a destination exists before naming it, and nothing more: no id, no wallet, no balance, no status. The lookup is exact after normalisation (a leading @ and upper case are accepted). Only consumer handles resolve here; a Business @banza answers 404. Credential: developer key with customers:read (no project binding required), or a merchant credential.",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "handle_xxx",
            "description": "The @banza, with or without the leading @"
          }
        ],
        "responses": {
          "200": {
            "description": "The handle exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandleResolution"
                },
                "example": {
                  "handle": "handle_xxx",
                  "display_name": "Example Customer"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "INSUFFICIENT_SCOPE — missing customers:read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No consumer holds this handle (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sandbox API key sent directly as the Bearer token (bz_test_sk_XXXX placeholder — never a real key in examples). bz_live_ keys are rejected fail-closed; there is no Production key issuance."
      },
      "statusToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "bzst_",
        "description": "A realtime status token (bzst_…) from PaymentSession.realtime.token. Read-only, bound to one Payment Session and one environment, valid for at most 30 minutes. Sent only in the Authorization header — a token in the URL is refused (400 REALTIME_TOKEN_IN_URL). It is not an API key and opens nothing else."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "example": "idem_xxx",
        "description": "Send on writes to retry safely. Sandbox behaviour: the original 2xx/4xx response is replayed for the same key for 24 hours (scoped per credential, method and path); 5xx responses are never replayed; concurrent requests with the same key receive 409 CONFLICT until the first completes."
      },
      "SessionId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "example": "payment_session_xxx"
      }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Canonical flat error envelope returned by every failing endpoint (observed Sandbox behaviour). Handle errors by code, never by message; keep request_id for support correlation.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable UPPER_SNAKE error code (e.g. VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, RATE_LIMITED, INTERNAL_ERROR)"
          },
          "message": {
            "type": "string",
            "description": "Human-readable message (not stable — do not branch on it)"
          },
          "request_id": {
            "type": "string",
            "description": "Correlation id for support (placeholder req_xxx in examples)"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CreatePaymentLinkRequest": {
        "type": "object",
        "properties": {
          "wallet_id": {
            "type": "string",
            "description": "Destination wallet"
          },
          "amount_minor": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Omit for an open-amount link"
          },
          "currency": {
            "type": "string",
            "example": "AOA"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "wallet_id",
          "currency"
        ]
      },
...
    }
  }
}