# services/api-gateway/internal/handler/payment_links.go
package handler

import (
	"context"
	"encoding/json"
	"errors"
	"github.com/google/uuid"
	"io"
	"net/http"
	"strconv"
	"time"

	"github.com/go-chi/chi/v5"

	"github.com/banzami/banzami/services/api-gateway/internal/apierror"
	"github.com/banzami/banzami/services/api-gateway/internal/middleware"
	"github.com/banzami/banzami/services/api-gateway/internal/service"
)

type PaymentLinkHandler struct {
	seal        bindingSealer
	wallets     walletReader
	svc         service.PaymentLinkService
	merchantSvc service.MerchantService
	webhookSvc  service.WebhookService
	identities  businessIdentityLookup
	realtime    *service.RealtimeTokens
	linkSession func(ctx context.Context, linkID string) (*service.PaymentSession, error)
}

// WithRealtime lets the public link view carry a status token for the Payment
// Session the link belongs to.
func (h *PaymentLinkHandler) WithRealtime(t *service.RealtimeTokens, sessions service.PaymentSessionService) *PaymentLinkHandler {
	if t == nil || sessions == nil {
		return h
	}
	h.realtime = t
	h.linkSession = func(ctx context.Context, linkID string) (*service.PaymentSession, error) {
		return sessions.GetByInterface(ctx, "link", linkID)
	}
	return h
}

// WithWallets gives the handler what it needs to check that a Business's link
// pays into its own wallet. Without it every merchant-created link is refused.
func (h *PaymentLinkHandler) WithWallets(w walletReader) *PaymentLinkHandler {
	h.wallets = w
	return h
}

func NewPaymentLinkHandler(svc service.PaymentLinkService, merchantSvc service.MerchantService, webhookSvc service.WebhookService) *PaymentLinkHandler {
	return &PaymentLinkHandler{svc: svc, merchantSvc: merchantSvc, webhookSvc: webhookSvc}
}

// WithBindingSeal supplies the ADR-055 seal used when a developer key ISSUES a
// payment link. Nil leaves the handler unable to issue one for a developer key,
// which is the correct fail-closed default rather than issuing an unsealed
// artifact.
func (h *PaymentLinkHandler) WithBindingSeal(s bindingSealer) *PaymentLinkHandler {
	h.seal = s
	return h
}

// POST /v1/payment-links
// merchantPrincipalID returns the merchant id of an authenticated merchant JWT.
//
// Every payment-link operation must be bound to it. The developer-key path was
// given tenant isolation when ADR-047 introduced it; the older merchant-JWT path
// never was, so a merchant could name any merchant_id in a body or query and the
// gateway believed it. Measured on the deployed Sandbox: merchant B created a
// link payable to merchant A, read A's private record, listed A's links, and
// CANCELLED A's link — the cancellation took effect, leaving A's payment link
// dead (RA-047).
func merchantPrincipalID(r *http.Request) (string, bool) {
	p, ok := middleware.GetPrincipal(r.Context())
	if !ok || p.MerchantID == "" {
		return "", false
	}
	return p.MerchantID, true
}

// linkActor is who may act on payment links: a bound developer key holding the
// scope (payee from its Project binding), or a merchant JWT. Nothing else.
//
// A consumer token is signed with the same HS256 secret and authenticates on
// this surface with no merchant id. Every check below keyed on "is there a
// merchant principal?", so a consumer token — even one naming a consumer that
// does not exist — skipped them all: measured on the deployed Sandbox, it
// listed another merchant's links and passed ownership on cancel. A developer
// key had the same gap on list, cancel and mark-used.
func linkActor(w http.ResponseWriter, r *http.Request, scope string) (string, bool) {
	payee, handled, isDev := developerPaymentAuthority(w, r, scope)
	if handled {
		return "", false
	}
	if isDev {
		return payee.merchantID, true
	}
	if mid, ok := merchantPrincipalID(r); ok {
		return mid, true
	}
	apierror.Respond(w, r, http.StatusUnauthorized, "UNAUTHORIZED",
		"merchant credentials or a project key are required")
	return "", false
}

// linkIDIsWellFormed rejects an id that cannot name a link, before any lookup.
//
// A malformed id is the caller's mistake, not the operator's failure. Passing it
// through reached a uuid parse in the data layer and surfaced as 500
// INTERNAL_ERROR — most easily hit by handing this route a payment-link SLUG,
// which is a natural confusion since the slug is what the public URL carries.
// Answered as not-found for the same reason a foreign link is: an id that cannot
// exist and one that does not exist are the same answer to the caller.
func linkIDIsWellFormed(w http.ResponseWriter, r *http.Request, id string) bool {
	if _, err := uuid.Parse(id); err != nil {
		apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
		return false
	}
	return true
}

// requireOwnedLink loads a link and confirms the caller owns it.
//
// A link belonging to someone else is reported as missing, not forbidden: 403
// would confirm the id exists and make the resource enumerable. This matches the
// privacy behaviour the payment-session surface already uses for a cross-merchant
// read, and the behaviour the developer-key path already had here.
func (h *PaymentLinkHandler) requireOwnedLink(w http.ResponseWriter, r *http.Request, id, scope string) (*service.PaymentLink, bool) {
	actor, ok := linkActor(w, r, scope)
	if !ok {
		return nil, false
	}
	if !linkIDIsWellFormed(w, r, id) {
		return nil, false
	}
	link, err := h.svc.Get(r.Context(), id)
	if err != nil {
		if errors.Is(err, service.ErrPaymentLinkNotFound) {
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
			return nil, false
		}
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not fetch payment link")
		return nil, false
	}
	if link.MerchantID != actor {
		apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
		return nil, false
	}
	return link, true
}

func (h *PaymentLinkHandler) Create(w http.ResponseWriter, r *http.Request) {
	// Developer-key authority (payee from the Project binding) OR merchant JWT
	// (existing body-supplied identity). Scope enforced before business logic.
	dev, handled, isDev := developerPaymentAuthorityForArtifact(w, r, "payment_links:write", h.seal)
	if handled {
		return
	}
	raw, err := io.ReadAll(io.LimitReader(r.Body, 1<<20))
	if err != nil {
		apierror.Respond(w, r, http.StatusBadRequest, "INVALID_BODY", "could not read request body")
		return
	}
	// A developer request may never carry a payee — it derives ONLY from the
	// Project binding.
	if isDev && rejectClientPayeeFields(w, r, raw) {
		return
	}
	var body struct {
		MerchantID  string     `json:"merchant_id"`
		WalletID    string     `json:"wallet_id"`
		AmountMinor *int64     `json:"amount_minor"`
		Currency    string     `json:"currency"`
		Description *string    `json:"description"`
		ExpiresAt   *time.Time `json:"expires_at"`
	}
	if len(raw) > 0 {
		if err := json.Unmarshal(raw, &body); err != nil {
			apierror.Respond(w, r, http.StatusBadRequest, "INVALID_BODY", "request body must be valid JSON")
			return
		}
	}

	req := service.CreatePaymentLinkRequest{
		AmountMinor: body.AmountMinor,
		Currency:    body.Currency,
		Description: body.Description,
		ExpiresAt:   body.ExpiresAt,
	}
	if isDev {
		// Payee derives exclusively from the binding.
		req.MerchantID = dev.merchantID
		req.WalletID = dev.walletID
		req.WalletAccountID = dev.walletAccountID
		req.SandboxProjectID = dev.projectID
	} else {
		// A merchant JWT may only create links payable to ITSELF. The body value
		// used to be trusted, so any merchant could mint a link collecting into
		// another merchant's wallet under that merchant's identity (RA-047).
		mid, ok := merchantPrincipalID(r)
		if !ok {
			apierror.Respond(w, r, http.StatusUnauthorized, "UNAUTHORIZED", "merchant authentication required")
			return
		}
		if body.MerchantID != "" && body.MerchantID != mid {
			apierror.Respond(w, r, http.StatusForbidden, "FORBIDDEN",
				"a payment link may only be created for the authenticated merchant")
			return
		}
		// And into its OWN wallet. The wallet was taken from the body with no
		// ownership check anywhere down to core, and acquiring credits the
		// link's wallet — so Business A could open a link collecting into
		// Business B's wallet. Answered as not found: no existence oracle.
		if body.WalletID != "" && !ownsWallet(r.Context(), h.wallets, body.WalletID, mid) {
			businessTenantDenials.WithLabelValues(tenantSurfacePaymentLink).Inc()
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "wallet not found")
			return
		}
		req.MerchantID = mid
		req.WalletID = body.WalletID
	}

	switch {
	case req.MerchantID == "":
		apierror.Respond(w, r, http.StatusBadRequest, "MISSING_FIELD", "merchant_id is required")
		return
	case req.WalletID == "":
		apierror.Respond(w, r, http.StatusBadRequest, "MISSING_FIELD", "wallet_id is required")
		return
	case body.Currency == "":
		apierror.Respond(w, r, http.StatusBadRequest, "MISSING_FIELD", "currency is required")
		return
	case body.AmountMinor != nil && *body.AmountMinor <= 0:
		apierror.Respond(w, r, http.StatusBadRequest, "INVALID_AMOUNT", "amount_minor must be positive")
		return
	case body.ExpiresAt != nil && body.ExpiresAt.Before(time.Now()):
		apierror.Respond(w, r, http.StatusBadRequest, "INVALID_EXPIRY", "expires_at must be in the future")
		return
	}

	link, err := h.svc.Create(r.Context(), req)
	if err != nil {
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not create payment link")
		return
	}
	respond(w, http.StatusCreated, link)
}

// GET /v1/payment-links
func (h *PaymentLinkHandler) List(w http.ResponseWriter, r *http.Request) {
	// The caller lists ITS OWN links — the merchant of its JWT or of its key's
	// binding. The query parameter used to select the tenant (RA-047), and for a
	// developer key or a consumer token it still did (see linkActor).
	merchantID, ok := linkActor(w, r, "payment_links:read")
	if !ok {
		return
	}
	if q := r.URL.Query().Get("merchant_id"); q != "" && q != merchantID {
		apierror.Respond(w, r, http.StatusForbidden, "FORBIDDEN",
			"payment links can only be listed for your own merchant")
		return
	}
	// Both are checked here, where the caller can be told what is wrong. An
	// unparsable cursor used to reach core, come back as a 400 and leave this
	// handler as a 500 — a caller's typo reported as the server's failure — and
	// a limit had no upper bound at all.
	limit := int64(20)
	if l := r.URL.Query().Get("limit"); l != "" {
		v, err := strconv.ParseInt(l, 10, 64)
		if err != nil || v < 1 || v > 100 {
			apierror.Respond(w, r, http.StatusBadRequest, "INVALID_PARAM", "limit must be between 1 and 100")
			return
		}
		limit = v
	}
	cursor := r.URL.Query().Get("cursor")
	if cursor != "" {
		if _, err := uuid.Parse(cursor); err != nil {
			apierror.Respond(w, r, http.StatusBadRequest, "INVALID_PARAM", "cursor must be the next_cursor of a previous page")
			return
		}
	}
	page, err := h.svc.List(r.Context(), service.ListPaymentLinksRequest{
		MerchantID: merchantID,
		Limit:      limit,
		Cursor:     cursor,
	})
	if err != nil {
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not list payment links")
		return
	}
	respond(w, http.StatusOK, page)
}

// GET /v1/payment-links/{id}
func (h *PaymentLinkHandler) Get(w http.ResponseWriter, r *http.Request) {
	// A developer key must hold payment_links:read and may only read links of its
	// bound merchant (tenant isolation) — enforced before the record is exposed.
	dev, handled, isDev := developerPaymentAuthority(w, r, "payment_links:read")
	if handled {
		return
	}
	if _, ok := merchantPrincipalID(r); !isDev && !ok {
		apierror.Respond(w, r, http.StatusUnauthorized, "UNAUTHORIZED",
			"merchant credentials or a project key are required")
		return
	}
	id := chi.URLParam(r, "id")
	if !linkIDIsWellFormed(w, r, id) {
		return
	}
	link, err := h.svc.Get(r.Context(), id)
	if err != nil {
		if errors.Is(err, service.ErrPaymentLinkNotFound) {
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
			return
		}
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not fetch payment link")
		return
	}
	if isDev && link.MerchantID != dev.merchantID {
		// Cross-tenant read is indistinguishable from missing.
		apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
		return
	}
	// The same isolation for a merchant JWT. Only the developer-key path had it,
	// so a merchant could read any link by id (RA-047).
	if mid, ok := merchantPrincipalID(r); ok && link.MerchantID != mid {
		apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
		return
	}
	// refund_source is a merchant-private field: surface it only to the owning
	// merchant JWT. Developer keys and non-owners never see it.
	if p, ok := middleware.GetPrincipal(r.Context()); !ok || p.MerchantID != link.MerchantID {
		link.RefundSource = nil
	}
	respond(w, http.StatusOK, link)
}

// DELETE /v1/payment-links/{id}  → cancel
func (h *PaymentLinkHandler) Cancel(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	// Ownership BEFORE mutation. This had no check at all, so any merchant could
	// cancel any other merchant's link — and it took effect, killing the victim's
	// ability to be paid through it (RA-047).
	if _, ok := h.requireOwnedLink(w, r, id, "payment_links:write"); !ok {
		return
	}
	link, err := h.svc.Cancel(r.Context(), id)
	if err != nil {
		switch {
		case errors.Is(err, service.ErrPaymentLinkNotFound):
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
		case errors.Is(err, service.ErrPaymentLinkNotActive):
			apierror.Respond(w, r, http.StatusUnprocessableEntity, "LINK_NOT_ACTIVE", "payment link is not active")
		default:
			apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not cancel payment link")
		}
		return
	}
	respond(w, http.StatusOK, link)
}

// POST /v1/payment-links/{id}/mark-used — retired.
//
// It marked a link USED and sent payment_link.paid with no money moving: an
// integration could be told, by its own call, that a link was paid when nobody
// had paid it; on a link that belongs to a Payment Session the session stayed
// ACTIVE beside a "paid" link (A4-08). A link is paid only by a payment, and
// core marks it and emits the event in that payment's own transaction. A link
// the Business wants closed is cancelled: DELETE /v1/payment-links/{id}.
func (h *PaymentLinkHandler) MarkUsed(w http.ResponseWriter, r *http.Request) {
	apierror.Respond(w, r, http.StatusGone, "ROUTE_RETIRED",
		"a payment link is marked paid only by a payment; to close an unpaid link, cancel it with DELETE /v1/payment-links/{id}")
}

// dispatchPaymentLinkPaid emits payment_link.paid to the merchant's registered
// endpoints.
//
// Every path that marks a link paid MUST go through here. This used to live
// inline in the merchant-facing MarkUsed handler only, which meant the event
// fired when a MERCHANT declared a link used but not when a PAYER actually paid
// it: the acquiring callback and the Sandbox confirm rail both mark the link
// used on the service directly. The practical effect was that the one webhook
// an integration actually depends on — "your customer paid" — never arrived,
// while a merchant-initiated mark-used did. Keeping the dispatch in one function
// is what stops that asymmetry coming back.
//
// Callers must have established authority over the link BEFORE calling this:
// emitting the event for someone else's link would deliver payment_link.paid to
// THEIR endpoints (RA-047).
//
// Fire-and-forget: the caller's response is not held up by delivery, which the
// WebhookService tracks and retries independently.
func dispatchPaymentLinkPaid(webhookSvc service.WebhookService, link *service.PaymentLink) {
	if webhookSvc == nil || link == nil {
		return
	}
	go func(l *service.PaymentLink) {
		payload, err := json.Marshal(l)
		if err != nil {
			return
		}
		_, _ = webhookSvc.Dispatch(context.Background(), service.DispatchRequest{
			MerchantID: l.MerchantID,
			EventType:  "payment_link.paid",
			Payload:    json.RawMessage(payload),
		})
	}(link)
}

// ---------------------------------------------------------------------------
// Public endpoints — no authentication required
// ---------------------------------------------------------------------------

// publicPaymentLink is the payer-safe view of a payment link (RT03 §4). It
// exposes ONLY what an unauthenticated payer needs and never leaks internal
// database identifiers (link/merchant/wallet UUIDs), the wallet account, or the
// operator refund_source. The payer POST derives merchant/wallet from the slug
// server-side, so these ids are not needed client-side.
//
// The payee is the Business's PUBLIC identity (business_public_identities): the
// name it presents and the @banza it owns. It used to be merchants.name — for a
// Business made by the retired Console setup that is "Sandbox · <Project>", the
// name of a Project that created the link, which is what pay.banzami.com showed
// the payer ("Sandbox · Doa-Sandbox", and no @doa).
//
// Paid says whether the payment this link asked for has been made. A link that
// belongs to a Payment Session is retired (CANCELLED) when the session is paid
// by its QR; the status alone told the payer opening the link that it was
// cancelled ("Link inválido") when the payment had in fact been made.
type publicPaymentLink struct {
	Slug           string     `json:"slug"`
	AmountMinor    *int64     `json:"amount_minor"`
	Currency       string     `json:"currency"`
	Description    *string    `json:"description"`
	Status         string     `json:"status"`
	Paid           bool       `json:"paid"`
	ExpiresAt      *time.Time `json:"expires_at"`
	PaidAt         *time.Time `json:"paid_at"`
	MerchantName   string     `json:"merchant_name"`
	MerchantHandle *string    `json:"merchant_handle"`
	// Realtime lets the hosted page watch the Payment Session this link belongs
	// to (ADR-060 §9): a single-session, read-only status token. Absent for a
	// link that belongs to no session. The slug already grants a public status
	// read, and the token grants no more than that.
	Realtime *publicRealtime `json:"realtime,omitempty"`
}

type publicRealtime struct {
	SessionID string `json:"session_id"`
	Token     string `json:"token"`
	ExpiresAt string `json:"expires_at"`
	Path      string `json:"path"`
}

func toPublicPaymentLink(link *service.PaymentLink, payee service.BusinessIdentity, paid bool) publicPaymentLink {
	view := publicPaymentLink{
		Slug:         link.Slug,
		AmountMinor:  link.AmountMinor,
		Currency:     link.Currency,
		Description:  link.Description,
		Status:       link.Status,
		Paid:         paid,
		ExpiresAt:    link.ExpiresAt,
		PaidAt:       link.PaidAt,
		MerchantName: payee.DisplayName,
	}
	if payee.Handle != "" {
		handle := payee.Handle
		view.MerchantHandle = &handle
	}
	return view
}

// businessIdentityLookup reads a Business's public identity.
type businessIdentityLookup func(ctx context.Context, merchantID string) (service.BusinessIdentity, error)

// WithBusinessIdentities gives the public payer view the Business's canonical
// public identity (business_public_identities, in the proofs database). Without
// it the view falls back to the account name — the defect this exists to
// remove — so wire it wherever the handler is built.
func (h *PaymentLinkHandler) WithBusinessIdentities(proofs *service.ProofService) *PaymentLinkHandler {
	if proofs == nil || proofs.Pool() == nil {
		return h
	}
	pool := proofs.Pool()
	h.identities = func(ctx context.Context, merchantID string) (service.BusinessIdentity, error) {
		return service.LookupBusinessIdentity(ctx, pool, merchantID)
	}
	return h
}

// publicPayee is the payee as the payer sees it. With the identity lookup wired
// a failed lookup names nobody, rather than falling back to the account name.
func (h *PaymentLinkHandler) publicPayee(ctx context.Context, merchantID string) (service.BusinessIdentity, error) {
	if h.identities != nil {
		b, err := h.identities(ctx, merchantID)
		if err != nil {
			return service.BusinessIdentity{}, nil
		}
		return b, nil
	}
	merchant, err := h.merchantSvc.Get(ctx, merchantID)
	if err != nil {
		return service.BusinessIdentity{}, err
	}
	return service.BusinessIdentity{DisplayName: merchant.Name}, nil
}

// GET /public/pay/{slug}
func (h *PaymentLinkHandler) GetPublic(w http.ResponseWriter, r *http.Request) {
	slug := chi.URLParam(r, "slug")
	link, err := h.svc.GetBySlug(r.Context(), slug)
	if err != nil {
		if errors.Is(err, service.ErrPaymentLinkNotFound) {
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
			return
		}
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not fetch payment link")
		return
	}

	payee, err := h.publicPayee(r.Context(), link.MerchantID)
	if err != nil {
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not fetch merchant")
		return
	}

	view := toPublicPaymentLink(link, payee, h.linkPaid(r.Context(), link))
	if h.realtime != nil && h.linkSession != nil {
		if sess, serr := h.linkSession(r.Context(), link.ID); serr == nil && sess != nil && sess.SessionID != "" {
			tok, exp := h.realtime.Mint(sess.SessionID)
			view.Realtime = &publicRealtime{SessionID: sess.SessionID, Token: tok, ExpiresAt: exp.Format(time.RFC3339),
				Path: "/v1/realtime/payment-sessions/" + sess.SessionID}
		}
	}
	respond(w, http.StatusOK, view)
}

// linkSessionStatus is implemented by a link service that can tell the status of
// the Payment Session a link belongs to ("" when it belongs to none).
type linkSessionStatus interface {
	SessionStatus(ctx context.Context, linkID string) (string, error)
}

// linkPaid reports whether the payment a link asked for has been made: the link
// was used, or it was retired because its Payment Session was paid through the
// session's other interface, the QR. Any failure to read the session answers
// "not known to be paid" — the link keeps its own status, never a claim of
// payment the operator could not confirm.
func (h *PaymentLinkHandler) linkPaid(ctx context.Context, link *service.PaymentLink) bool {
	if link.Status == "USED" {
		return true
	}
	if link.Status != "CANCELLED" {
		return false
	}
	sessions, ok := h.svc.(linkSessionStatus)
	if !ok {
		return false
	}
	status, err := sessions.SessionStatus(ctx, link.ID)
	return err == nil && status == "PAID"
}

// GET /public/pay/{slug}/status
func (h *PaymentLinkHandler) Status(w http.ResponseWriter, r *http.Request) {
	slug := chi.URLParam(r, "slug")
	link, err := h.svc.GetBySlug(r.Context(), slug)
	if err != nil {
		if errors.Is(err, service.ErrPaymentLinkNotFound) {
			apierror.Respond(w, r, http.StatusNotFound, "NOT_FOUND", "payment link not found")
			return
		}
		apierror.Respond(w, r, http.StatusInternalServerError, "INTERNAL_ERROR", "could not fetch payment link")
		return
	}
	respond(w, http.StatusOK, map[string]bool{"paid": h.linkPaid(r.Context(), link)})
}